# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "cell-analysis-backend",
        "description": "Backend API for cancer cell analysis application, apoptosis and viable cell",
        "contact": {
            "name": "Your Name",
            "email": "your.email@example.com"
        },
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/folders": {
            "get": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "List all folders for the authenticated user",
                "operationId": "list_folders",
                "responses": {
                    "200": {
                        "description": "List of folders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_FolderListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Create a new folder",
                "operationId": "create_folder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFolderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Folder created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_FolderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/folders/{folder_id}": {
            "delete": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Delete a folder and all its images (cascade delete)",
                "operationId": "delete_folder",
                "parameters": [
                    {
                        "name": "folder_id",
                        "in": "path",
                        "description": "Folder ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_DeleteFolderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Folder not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Rename a folder",
                "operationId": "rename_folder",
                "parameters": [
                    {
                        "name": "folder_id",
                        "in": "path",
                        "description": "Folder ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFolderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Folder renamed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_FolderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Folder not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "type": "object",
                "description": "API error structure",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "ApiResponse_DeleteFolderResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Delete folder response",
                        "required": [
                            "message",
                            "deleted_images_count"
                        ],
                        "properties": {
                            "deleted_images_count": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "message": {
                                "type": "string"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiResponse_FolderListResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "List folders response",
                        "required": [
                            "folders",
                            "total"
                        ],
                        "properties": {
                            "folders": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/FolderResponse"
                                }
                            },
                            "total": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponse_FolderResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Single folder response",
                        "required": [
                            "folder_id",
                            "folder_name",
                            "image_count",
                            "created_at"
                        ],
                        "properties": {
                            "created_at": {
                                "type": "string"
                            },
                            "deleted_at": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "folder_id": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "folder_name": {
                                "type": "string"
                            },
                            "image_count": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "CreateFolderRequest": {
                "type": "object",
                "description": "Create folder request",
                "required": [
                    "folder_name"
                ],
                "properties": {
                    "folder_name": {
                        "type": "string"
                    }
                }
            },
...
            "FolderResponse": {
                "type": "object",
                "description": "Single folder response",
                "required": [
                    "folder_id",
                    "folder_name",
                    "image_count",
                    "created_at"
                ],
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "folder_id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "folder_name": {
                        "type": "string"
                    },
                    "image_count": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "UpdateFolderRequest": {
                "type": "object",
                "description": "Update folder request (rename)",
                "required": [
                    "folder_name"
                ],
                "properties": {
                    "folder_name": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}